docs: add JSDoc comments to improve code documentation
- Add JSDoc comments to src/lib/store/preferences.ts functions
- Add JSDoc comments to src/components/ui/sonner.tsx
- Add JSDoc comments to src/components/profile/profile-page.tsx
- Add JSDoc comments to src/components/settings/email-change-form.tsx

All functions now have proper JSDoc documentation explaining their purpose, parameters, and return values where applicable.
