+++ package.json
@@ -65,6 +65,7 @@
     "@tanstack/react-router": "^1.168.10",
     "@tanstack/react-router-devtools": "^1.166.11",
     "@tanstack/react-start": "^1.167.16",
+    "@tanstack/react-store": "^0.9.3",
     "a11yer": "^0.1.0",
     "better-auth": "^1.5.6",
     "class-variance-authority": "^0.7.1",
@@ -82,6 +83,7 @@
     "remark-gfm": "^4.0.1",
     "shadcn": "^4.1.2",
     "shiki": "^4.0.2",
+    "sonner": "^2.0.7",
     "tailwind-merge": "^3.5.0",
     "tw-animate-css": "^1.4.0",
     "zod": "^4.3.6"
@@ -103,6 +105,7 @@
     "@typescript/native-preview": "^7.0.0-dev.20260403.1",
     "@vitejs/plugin-react": "^6.0.1",
     "actions-up": "^1.12.0",
+    "better-sqlite3": "^12.8.0",
     "drizzle-kit": "^0.31.10",
     "jsdom": "^29.0.1",
     "k6": "^0.0.0",
