+++ src/components/ui/sonner.tsx
@@ -1,0 +1,44 @@
+/**
+ * Custom Toaster component that wraps Sonner with theme support.
+ * Provides consistent toast styling and theming based on application theme.
+ */
+import { CircleCheck, Info, LoaderCircle, OctagonX, TriangleAlert } from "lucide-react";
+import { Toaster as Sonner } from "sonner";
+import { useTheme } from "../theme/provider";
+
+type ToasterProps = React.ComponentProps<typeof Sonner>;
+
+/**
+ * Toaster component that integrates with the application's theme system.
+ * @param props - Props passed to the underlying Sonner Toaster component
+ * @returns JSX.Element Themed Toaster component
+ */
+const Toaster = ({ ...props }: ToasterProps) => {
+  const { theme = "system" } = useTheme();
+
+  return (
+    <Sonner
+      theme={theme as ToasterProps["theme"]}
+      className="toaster group"
+      icons={{
+        success: <CircleCheck className="h-4 w-4" />,
+        info: <Info className="h-4 w-4" />,
+        warning: <TriangleAlert className="h-4 w-4" />,
+        error: <OctagonX className="h-4 w-4" />,
+        loading: <LoaderCircle className="h-4 w-4 animate-spin" />,
+      }}
+      toastOptions={{
+        classNames: {
+          toast:
+            "group toast group-[.toaster]:bg-background group-[.toaster]:text-foreground group-[.toaster]:border-border group-[.toaster]:shadow-lg",
+          description: "group-[.toast]:text-muted-foreground",
+          actionButton: "group-[.toast]:bg-primary group-[.toast]:text-primary-foreground",
+          cancelButton: "group-[.toast]:bg-muted group-[.toast]:text-muted-foreground",
+        },
+      }}
+      {...props}
+    />
+  );
+};
+
+export { Toaster };
