+++ test/setup.ts
@@ -1,6 +1,30 @@
 import { JSDOM } from "jsdom";
 import { HOST, isBrowser, PORT } from "../src/config";
 
+// Mock localStorage for tests
+const localStorageMock: Storage = (() => {
+  let store: Record<string, string> = {};
+  return {
+    getItem: (key: string) => store[key] || null,
+    setItem: (key: string, value: string) => {
+      store[key] = value.toString();
+    },
+    removeItem: (key: string) => {
+      delete store[key];
+    },
+    clear: () => {
+      store = {};
+    },
+    length: 0,
+    key: (index: number) => {
+      const keys = Object.keys(store);
+      return index >= 0 && index < keys.length ? keys[index] : null;
+    },
+  };
+})();
+
+Object.defineProperty(globalThis, "localStorage", { value: localStorageMock });
+
 const url =
   import.meta.env.VITE_API_URL || (isBrowser ? window.location.origin : `http://${HOST}:${PORT}`);
